pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-backend-api>=1.8.0",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "reflex>=0.7.0",
    "structlog>=25.1.0",
]

[dependency-groups]
dev = ["pytest>=8.3.4", "ruff>=0.9.4"]
